fix: 현재 연월의 부동산 정보 수집 스케줄러 개발 중, poetry와 c 라이브러리 의존성 때문에 dockerfile 수정

+++ real_estate/management/commands/insert_real_estates_for_searching_command.py
@@ -145,7 +145,8 @@ class Command(BaseCommand):
         real_estate_deal4.save()
 
     def handle(self, *args, **options):
-        self.delete_real_estate_deal()
-        self.delete_real_estate()
-        self.create_real_estate_and_real_estate_deal()
-        self.run_indexer()
+        if os.getenv("SERVER_ENV") in ["local"]:
+            self.delete_real_estate_deal()
+            self.delete_real_estate()
+            self.create_real_estate_and_real_estate_deal()
+            self.run_indexer()
